GOT-12: Estrutura para goals

+++ go-tracker-api/src/main/kotlin/go/tracker/api/request/trainer/MedalStatusRequest.kt
@@ -4,6 +4,7 @@ import go.tracker.api.common.validation.constraint.annotation.ValidEnumList
 import go.tracker.models.enums.Medals
 import go.tracker.models.trainer.TrainerMedalStatus
 import io.swagger.v3.oas.annotations.media.Schema
+import java.math.BigDecimal
 
 data class MedalStatusRequest(
     @field:ValidEnumList(enumClass = Medals::class, message = "Invalid medal status")
@@ -11,7 +12,7 @@ data class MedalStatusRequest(
     val medal: Medals? = null,
 
     @field: Schema(requiredMode = Schema.RequiredMode.REQUIRED)
-    val value: Long? = null,
+    val value: BigDecimal? = null,
 ) {
     fun toDomain(request: MedalStatusRequest): TrainerMedalStatus =
         TrainerMedalStatus().apply {
