+++ go-tracker-models/src/main/kotlin/go/tracker/models/trainer/medals/FiniteMedalsValues.kt
@@ -1,7 +1,9 @@
 package go.tracker.models.trainer.medals
 
-data class FiniteMedalsValues (
-    var unown: Long? = null,
-    var megaEvolutionGuru: Long? = null,
-    var vivillon: Long? = null,
+import java.math.BigDecimal
+
+data class FiniteMedalsValues(
+    var unown: BigDecimal? = null,
+    var megaEvolutionGuru: BigDecimal? = null,
+    var vivillon: BigDecimal? = null,
 )
